% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_contents.R
\name{mocis_contents}
\alias{mocis_contents}
\title{Extracts table of contents of a MoCiS object and places in data frame}
\usage{
mocis_contents(mocis_object)
}
\arguments{
\item{mocis_object}{A list returned from \code{MoCiS::mocis}}
}
\value{
A tibble with columns \code{loc} (locationm), \code{var} (variable) and \code{gen} (genus) and one row for each combination found in \code{mocis_object}
}
\description{
Extracts table of contents of a MoCiS object and places in data frame
}
