% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_name_species.R
\name{mocis_name_species}
\alias{mocis_name_species}
\title{Translates genus from short to long format}
\usage{
mocis_name_species(gen)
}
\arguments{
\item{gen, }{genus short form (e.g. CLUP)}
}
\value{
Genus long form (e.g. Herring)
}
\description{
Translates genus from short to long format
}
