% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_name_station.R
\name{mocis_name_station}
\alias{mocis_name_station}
\title{Translates station from short to long format}
\usage{
mocis_name_station(loc)
}
\arguments{
\item{loc, }{station short form (e.g. ABBE)}
}
\value{
Station long form (e.g. Abbekås)
}
\description{
Translates station from short to long format
}
