% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_report.R
\name{mocis_report}
\alias{mocis_report}
\title{Builds an RMarkdown  file from a mocis_object}
\usage{
mocis_report(mocis_object, file = "report.Rmd", contaminants,
  report_year)
}
\arguments{
\item{mocis_object, }{mocis_object A list returned from \code{MoCiS::mocis}}

\item{file, }{file name/path for output}

\item{contaminants, }{name of group of contaminants for chapter header (e.g. "metals")}

\item{report_year, }{year of report}
}
\description{
Builds an RMarkdown  file from a mocis_object
}
