% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_table.R
\name{mocis_table}
\alias{mocis_table}
\title{Builds a table of summary statistics}
\usage{
mocis_table(data, dec = 2)
}
\arguments{
\item{mocis_object}{A list returned from \code{MoCiS::mocis}}
}
\value{
A tibble with
}
\description{
Builds a table of summary statistics
}
