% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_tibble.R
\name{mocis_tibble}
\alias{mocis_tibble}
\title{Places the contents of a mocis_object in a tibble}
\usage{
mocis_tibble(mocis_object)
}
\arguments{
\item{mocis_object}{A list returned from \code{MoCiS::mocis}}
}
\value{
A tibble with
}
\description{
Places the contents of a mocis_object in a tibble
}
